+++ src/main/java/com/utp/trabajo/services/TipoProductoService.java
@@ -4,8 +4,10 @@
  */
 package com.utp.trabajo.services;
 
+import com.utp.trabajo.exception.security.NotEnoughPermissionsException;
 import com.utp.trabajo.model.dao.TipoProductoDao;
 import com.utp.trabajo.model.entities.TipoProducto;
+import com.utp.trabajo.services.security.SecurityService;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -20,9 +22,16 @@ public class TipoProductoService {
 
     @Autowired
     private TipoProductoDao tipoproductoDao;
+    
+    @Autowired
+    private SecurityService securityService;
 
     @Transactional(readOnly = true)
-    public List<TipoProducto> streamTipoProducto(Long lastId, Long limit) {
+    public List<TipoProducto> streamTipoProducto(Long lastId, Long limit) throws NotEnoughPermissionsException {
+        if (!securityService.getPermissions().contains("read")) {
+            throw new NotEnoughPermissionsException("Sin permisos de lectura.");
+        }
+        
         try ( Stream<TipoProducto> streamedTipoProducto = tipoproductoDao.findByIdTipoProductoGreaterThan(lastId)) {
             return streamedTipoProducto.limit(limit)
                 .collect(Collectors.toList());
@@ -30,23 +39,35 @@ public class TipoProductoService {
     }
 
     @Transactional(readOnly = true)
-    public TipoProducto encontrarTipoProductoPorId(Long idMarca) {
+    public TipoProducto encontrarTipoProductoPorId(Long idMarca) throws NotEnoughPermissionsException {
+        if (!securityService.getPermissions().contains("read")) {
+            throw new NotEnoughPermissionsException("Sin permisos de lectura.");
+        }
+        
         return tipoproductoDao.findById(idMarca).orElseThrow();
     }
 
     @Transactional
-    public void nuevoTipoProducto(TipoProducto tipoproducto) {
+    public void nuevoTipoProducto(TipoProducto tipoproducto) throws NotEnoughPermissionsException {
+        if (!securityService.getPermissions().contains("create")) {
+            throw new NotEnoughPermissionsException("Sin permisos de creación.");
+        }
         tipoproductoDao.save(tipoproducto);
 
     }
 
     @Transactional
-    public void actualizarTipoProducto(TipoProducto tipoproducto) {
-
+    public void actualizarTipoProducto(TipoProducto tipoproducto) throws NotEnoughPermissionsException {
+        if (!securityService.getPermissions().contains("edit")) {
+            throw new NotEnoughPermissionsException("Sin permisos de edición.");
+        }
     }
 
     @Transactional
-    public List<TipoProducto> eliminarTipoProducto(List<Long> idsTipoProducto) {
+    public List<TipoProducto> eliminarTipoProducto(List<Long> idsTipoProducto) throws NotEnoughPermissionsException {
+        if (!securityService.getPermissions().contains("delete")) {
+            throw new NotEnoughPermissionsException("Sin permisos de eliminación.");
+        }
         return tipoproductoDao.removeAllByIdTipoProductoIn(idsTipoProducto);
     }
 }
